Move helper functions to a separate file
